extract scopelistService
