+++ src/services/user-group/index.js
@@ -8,8 +8,8 @@ const {
 const hooks = require('./hooks');
 const courseGroupsHooks = require('./hooks/courseGroups');
 const courseCopyService = require('./course-copy-service');
+const courseScopelistService = require('./services/courseScopeLists');
 const classHooks = require('./hooks/classes');
-const { /* ScopePermissionService, */ ScopeListService } = require('../helpers/scopePermissions');
 
 // eslint-disable-next-line func-names
 module.exports = function () {
@@ -65,46 +65,5 @@ module.exports = function () {
 	const gradeService = app.service('/grades');
 	gradeService.hooks(hooks);
 
-	ScopeListService.initialize(app, '/users/:scopeId/courses', async (user, permissions, params) => {
-		let filter = 'active';
-		let substitution = 'false';
-		if (params.query.filter && ['active', 'archived', 'all'].includes(params.query.filter)) {
-			({ filter } = params.query);
-		}
-		if (params.query.substitution && ['true', 'false', 'all'].includes(params.query.substitution)) {
-			({ substitution } = params.query);
-		}
-
-		const userQuery = { $or: [] };
-		if (['false', 'all'].includes(substitution)) {
-			userQuery.$or.push(
-				{ userIds: user._id },
-				{ teacherIds: user._id },
-			);
-		}
-		if (['true', 'all'].includes(substitution)) userQuery.$or.push({ substitutionIds: user._id });
-
-		let untilQuery = {};
-		if (filter === 'active') {
-			untilQuery = {
-				$or: [
-					{ untilDate: { $exists: false } },
-					{ untilDate: { $gte: Date.now() } },
-				],
-			};
-		}
-		if (filter === 'archived') {
-			untilQuery = { untilDate: { $lt: Date.now() } };
-		}
-
-		return app.service('courses').find({
-			query: {
-				$and: [
-					userQuery,
-					untilQuery,
-				],
-			},
-			paginate: params.paginate,
-		});
-	});
+	app.configure(courseScopelistService);
 };
